Added a uselocalstorage hook
