+++ src/Styles/suggestionCard.module.css
@@ -16,6 +16,7 @@
     border-radius: 8px;
     box-shadow: var(--box-shadow);
     transition: box-shadow 0.3s ease;
+    transition: 0.6s ease-in-out;
 }
 
 .card h3 {
